+++ designer/server/src/test/scala/pl/touk/nussknacker/ui/validation/UIProcessValidatorSpec.scala
@@ -339,9 +339,10 @@ class UIProcessValidatorSpec extends AnyFunSuite with Matchers with TableDrivenP
       createGraph(
         List(
           Source("in", SourceRef(ProcessTestData.existingSourceFactory, List())),
+          Variable("var", "varName", "#meta.processLabels".spel),
           Sink("out", SinkRef(ProcessTestData.existingSinkFactory, List()))
         ),
-        List(Edge("in", "out", None)),
+        List(Edge("in", "var", None), Edge("var", "out", None)),
       )
     }