+++ docs/Changelog.md
@@ -54,6 +54,7 @@
   * Jackson 2.15.4 -> 2.17.2
   * cats 2.10 -> 2.12
 * [#6805](https://github.com/TouK/nussknacker/pull/6805) Support for Flink 1.19.1
+* [#6935](https://github.com/TouK/nussknacker/pull/6935) Spel: Scenario labels added to meta variable - `#meta.processLabels`
 
 ## 1.17